adding to readme and locked down resolvers.

+++ server/src/resolvers.ts
@@ -1,8 +1,7 @@
 import { Resolvers } from './generated/graphql'
 import { v4 as uuidv4 } from 'uuid'
 import { ApolloError } from 'apollo-server'
-import { isValidPassword, hashPassword, createJWT } from './utils'
-import { hash } from 'bcrypt'
+import { isValidPassword, hashPassword, createJWT, verifyJwt } from './utils'
 
 // Resolvers define the technique for fetching the types defined in the
 // schema.
@@ -11,13 +10,10 @@ import { hash } from 'bcrypt'
 // including the field name, the path to the field from the root, and more.
 export const resolvers: Resolvers = {
     Query: {
-        carsByUserID: (
-            parent,
-            args,
-            { user, db: { sequelize, User, Car } },
-            info
-        ) => {
+        carsByUserID: (parent, args, { user, db: { User, Car } }, info) => {
             try {
+                if (!user)
+                    throw new Error('Invalid action. User not logged in.')
                 return Car.findAll({
                     order: [['createdAt', 'DESC']],
                     include: [{ model: User }],
@@ -50,8 +46,10 @@ export const resolvers: Resolvers = {
         },
     },
     Mutation: {
-        createCar: (parent, { data }, { db: { Car } }, info) => {
+        createCar: (parent, { data }, { user, db: { Car } }, info) => {
             try {
+                if (!user)
+                    throw new Error('Invalid action. User not logged in.')
                 const id = uuidv4()
                 const { ...args } = data
                 return Car.create({ id, ...args })
@@ -62,26 +60,29 @@ export const resolvers: Resolvers = {
         updateCar: (
             parent,
             { data: { id, ...rest } },
-            { db: { Car } },
+            { user, db: { Car } },
             info
         ) => {
             try {
+                if (!user)
+                    throw new Error('Invalid action. User not logged in.')
                 return Car.update({ ...rest }, { where: { id } })
             } catch (err) {
                 throw new ApolloError('Could not update car.')
             }
         },
-        deleteCar: async (parent, { id }, { db: { Car } }, info) => {
+        deleteCar: async (parent, { id }, { user, db: { Car } }, info) => {
             try {
+                if (!user)
+                    throw new Error('Invalid action. User not logged in.')
                 await Car.destroy({ where: { id } })
                 return id
             } catch (err) {
-                throw new ApolloError('Could not delete car.')
+                throw new ApolloError(err)
             }
         },
         createUser: async (parent, { data }, { db: { User } }, info) => {
             try {
-                console.log('starting create user')
                 const { password, ...userInfo } = data
                 const prevUser = await User.findOne({
                     where: { username: data.username },
@@ -92,13 +93,11 @@ export const resolvers: Resolvers = {
                     )
 
                 const hashedPassword = await hashPassword(password)
-                console.log(hashedPassword)
                 const newUser = await User.create({
                     id: uuidv4(),
                     password: hashedPassword,
                     ...userInfo,
                 })
-                console.log(newUser)
 
                 return {
                     ...createJWT(newUser.username, newUser.id, newUser.email),
